refactor(companies): rename module folder path to companies
